+++ example/base_example.dart
@@ -15,7 +15,8 @@ Actor actorFactory(Uri path) {
     calls++;
 
     // just print the received message
-    print("actor ${context.current.path} received message: $message, call: $calls");
+    print(
+        "actor ${context.current.path} received message: $message, correlationId: ${context.correlationId}, call: $calls");
   };
 }