+++ lib/src/cluster/protocol.dart
@@ -53,7 +53,8 @@ class Protocol {
   final Duration _timeout;
   final Future<CreateActorResponse> Function(Uri path, int? mailboxSize, bool? useExistingActor) _handleCreateActor;
   final Future<LookupActorResponse> Function(Uri path) _handleLookupActor;
-  final Future<SendMessageResponse> Function(Uri path, Object? message, Uri? sender, Uri? replyTo) _handleSendMessage;
+  final Future<SendMessageResponse> Function(
+      Uri path, Object? message, Uri? sender, Uri? replyTo, String? correlationId) _handleSendMessage;
 
   Protocol(
     String id,
@@ -107,14 +108,15 @@ class Protocol {
     return completer.future;
   }
 
-  Future<void> sendMessage(Uri path, Object? message, Uri? sender, Uri? replyTo) {
-    _log.fine('sendMessage < path=$path, message=$message, sender=$sender, replyTo=$replyTo');
+  Future<void> sendMessage(Uri path, Object? message, Uri? sender, Uri? replyTo, String? correlationId) {
+    _log.fine(
+        'sendMessage < path=$path, message=$message, sender=$sender, replyTo=$replyTo, correlationId=$correlationId');
 
     final request = ProtocolMessage(
       ProtocolMessageType.request,
       sendMessageMessageName,
       Uuid().v4(),
-      SendMessageRequest(path, message, sender, replyTo, _serDes),
+      SendMessageRequest(path, message, sender, replyTo, correlationId, _serDes),
     );
 
     final completer = Completer<void>();
@@ -164,7 +166,13 @@ class Protocol {
               ProtocolMessageType.response,
               message.name,
               message.correlationId,
-              await _handleSendMessage(request.path, request.message, request.sender, request.replyTo),
+              await _handleSendMessage(
+                request.path,
+                request.message,
+                request.sender,
+                request.replyTo,
+                request.correlationId,
+              ),
             );
             _channel.sink.add(response);
             break;